studio 2

+++ datatypes/datatypes-studio/src/main/java/org/launchcode/Area.java
@@ -5,7 +5,7 @@ public class Area {
     public static void main(String[] args) {
         Scanner input = new Scanner(System.in);
         System.out.println("Enter a Radius:");
-        double radius = input.nextInt();
+        double radius = input.nextDouble();
         double area = Circle.getArea(radius);
         System.out.println("The area of a circle of a radius of " + radius + " is " + area);
